aider: Ensure proper PEM formatting and add debug output for public key in validate_google_token function.
# Aider chat conversation:

USER: I ran this command:

 ./social-test.sh

And got this output:

ERROR:  ValueError: Could not deserialize key data. The data may be in an incorrect format or it may be encrypted with an unsupported algorithm.
CONTEXT:  Traceback (most recent call last):
  PL/Python function "social", line 30, in <module>
    return validate_google_token(token)
  PL/Python function "social", line 24, in validate_google_token
    id_info = jwt.decode(token, public_key, algorithms=['RS256'], audience=google_client_id)
  PL/Python function "social", line 118, in decode
  PL/Python function "social", line 89, in decode_complete
  PL/Python function "social", line 151, in decode_complete
  PL/Python function "social", line 235, in _verify_signature
  PL/Python function "social", line 258, in prepare_key
  PL/Python function "social", line 24, in load_pem_public_key
  PL/Python function "social", line 1247, in load_pem_public_key
  PL/Python function "social", line 1511, in _handle_key_loading_error
PL/Python function "social"

ASSISTANT: The error message suggests that the key data is still not being deserialized correctly. This could be due to a variety of reasons, including incorrect PEM formatting or an issue with the key data itself. To address this, we need to ensure that the PEM key is correctly formatted and that we are using the correct key data.

Let's review and correct the PEM key formatting in the `validate_google_token` function within the `social.FUNCTION.sql` file. We will also add some debugging output to help diagnose the issue.

Here's the diff with the necessary changes:

```diff
--- sql/schema/social.FUNCTION.sql
+++ sql/schema/social.FUNCTION.sql
@@ ... @@
 def validate_google_token(token):
     # Get Google public keys
     response = requests.get('https://www.googleapis.com/oauth2/v1/certs')
     if response.status_code != 200:
         raise Exception('Failed to retrieve Google public keys')
     keys = response.json()
     # Decode the JWT header to find the 'kid' value
     headers = jwt.get_unverified_header(token)
     kid = headers['kid']
     # Find the key with the matching 'kid' and construct a PEM formatted key
     if kid not in keys:
         raise Exception('Key ID not found in Google public keys')
-    # Ensure the key is wrapped at 64 characters to conform to PEM format
-    key_body = keys[kid]
-    wrapped_key_body = '\n'.join([key_body[i:i+64] for i in range(0, len(key_body), 64)])
-    public_key = "-----BEGIN PUBLIC KEY-----\n{}\n-----END PUBLIC KEY-----".format(wrapped_key_body)
+    # Ensure the key is properly PEM formatted
+    public_key = keys[kid]
+    if not public_key.startswith("-----BEGIN PUBLIC KEY-----"):
+        public_key = "-----BEGIN PUBLIC KEY-----\n" + public_key
+    if not public_key.endswith("-----END PUBLIC KEY-----"):
+        public_key = public_key + "\n-----END PUBLIC KEY-----"
+    # Debug: Output the public key to verify its format
+    plpy.notice("Public key:\n{}".format(public_key))
     # Decode the token
     # Obtain the Google client ID set in the database configuration
     google_client_id = plpy.execute("SHOW app.google_client_id")[0]['app.google_client_id']
     id_info = jwt.decode(token, public_key, algorithms=['RS256'], audience=google_client_id)
     # If the token is valid, return the user's ID
     if id_info['iss'] in ['accounts.google.com', 'https://accounts.google.com']:
         return id_info['sub']
     else:
         raise jwt.InvalidIssuerError
 return validate_google_token(token)
```

This change ensures that the public key is properly formatted with the correct PEM headers and footers. Additionally, it adds a debug statement to output the public key, which can help in diagnosing the issue. Apply this patch to your `social.FUNCTION.sql` file, and then run the `social-test.sh` script again to see if the issue is resolved or to get more information about the public key format.